requirements.txt
streamlit
pymupdf
faiss-cpu
sentence-transformers
transformers
torch
